# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Open Pectus Aggregator",
    "contact": {
      "name": "Open Pectus",
      "url": "https://github.com/Open-Pectus/Open-Pectus"
    },
    "version": "0.1.dev"
  },
  "paths": {
...
    "/webpush/subscribe": {
      "post": {
        "tags": [
          "webpush"
        ],
        "summary": "Subscribe User",
        "operationId": "subscribe_user",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "x-identity",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Identity"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebPushSubscription"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WebPushKeys": {
        "properties": {
          "auth": {
            "type": "string",
            "title": "Auth"
          },
          "p256dh": {
            "type": "string",
            "title": "P256Dh"
          }
        },
        "type": "object",
        "required": [
          "auth",
          "p256dh"
        ],
        "title": "WebPushKeys"
      },
      "WebPushSubscription": {
        "properties": {
          "endpoint": {
            "type": "string",
            "minLength": 1,
            "format": "uri",
            "title": "Endpoint"
          },
          "keys": {
            "$ref": "#/components/schemas/WebPushKeys"
          }
        },
        "type": "object",
        "required": [
          "endpoint",
          "keys"
        ],
        "title": "WebPushSubscription"
      }
    }
  }
}